Add files via upload
This project creates and hosts a React app using Amazon S3. I built the app locally, tested it on my machine, then uploaded the build files to an S3 bucket. The bucket was configured for static website hosting so I could verify the app runs the same in the browser as it did on localhost. This shows how to use S3 as a simple hosting option for front-end applications.